:truck: move ui files (#72)
