+++ src/pages/play/mode/index.css.ts
@@ -4,8 +4,6 @@ import { COLORED_ICON_BUTTON_SIZE_PIXEL } from '@/components/button/constants';
 
 import { MYRANKINGITEMBOX_HEIGHT } from '@/pages/ranking/constants';
 
-import { flexOptions, sprinkles } from '@/shared/styles';
-
 export const wrapper = style([
   {
     display: 'grid',
@@ -25,25 +23,4 @@ export const myRank = style([
   {
     gridColumn: '1 / 3',
   },
-]);
-
-export const matchingModalWrapper = style([
-  flexOptions({ option: 'columnCenter' }),
-]);
-
-export const matchingMessage = style([
-  flexOptions({ option: 'row' }),
-  sprinkles({
-    textSize: '2x',
-    justifyContent: 'flexStart',
-    width: 'full',
-    marginBottom: '1x',
-    paddingLeft: '4x',
-  }),
-]);
-
-export const matchingImageWrapper = style([
-  sprinkles({
-    marginY: '2x',
-  }),
 ]);
