+++ src/pages/play/mode/ui/index.css.ts
@@ -1,0 +1,24 @@
+import { style } from '@vanilla-extract/css';
+
+import { flexOptions, sprinkles } from '@/shared/styles';
+
+export const matchingModalWrapper = style([
+  flexOptions({ option: 'columnCenter' }),
+]);
+
+export const matchingMessage = style([
+  flexOptions({ option: 'row' }),
+  sprinkles({
+    textSize: '2x',
+    justifyContent: 'flexStart',
+    width: 'full',
+    marginBottom: '1x',
+    paddingLeft: '4x',
+  }),
+]);
+
+export const matchingImageWrapper = style([
+  sprinkles({
+    marginY: '2x',
+  }),
+]);
